src/shared_retrieval/pinecone_retriever.py
# src/shared_retrieval/pinecone_retriever.py
"""Pinecone retriever implementation."""
import os
from typing import List

from langchain_core.documents import Document
from pinecone import Pinecone

from src.shared_retrieval.base import BaseRetriever


class PineconeRetriever(BaseRetriever):
    """Pinecone retriever."""

    def __init__(self):
        """Initialize the Pinecone retriever."""
        pinecone_api_key = os.getenv("PINECONE_API_KEY")
        pinecone_index_name = os.getenv("PINECONE_INDEX_NAME")
        self.client = Pinecone(api_key=pinecone_api_key)
        self.index = self.client.Index(pinecone_index_name)

    def get_relevant_documents(self, query: str) -> List[Document]:
        """Get relevant documents based on a query."""
        # Add Pinecone specific logic here
        # Retrieving from Pinecone for query
        # Placeholder for actual Pinecone query
        return []
