pyproject.toml


[project]
name = "langgraph_eng_package"
version = "0.0.1"
dependencies = [
  "langgraph>=0.2.6",
  "langchain-openai>=0.1.22",
  "langchain-anthropic>=0.1.23",
  "langchain>=0.2.14",
  "langchain-fireworks>=0.1.7",
  "python-dotenv>=1.0.1",
  "langchain-elasticsearch>=0.2.2,<0.3.0",
  "langchain-pinecone>=0.1.3,<0.2.0",
  "msgspec>=0.18.6",
  "langchain-mongodb>=0.1.9",
  "langchain-cohere>=0.2.4",
  "google-generativeai",
  "groq",
  "langchain-community",
  "deepseek",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1", "pytest-asyncio", "pytest"]

[tool.setuptools]
packages = ["enrichment_agent", "engineer_agent", "retrieval_graph", "index_graph", "shared"]
[tool.setuptools.package-dir]
"enrichment_agent" = "src/enrichment_agent"
"engineer_agent" = "src/engineer_agent"
"langgraph.templates.enrichment_agent" = "src/enrichment_agent"
"langgraph.templates.retrieval_graph" = "src/rag_agents/retrieval_graph"
"langgraph.templates.index_graph" = "src/rag_agents/index_graph"
"retrieval_graph" = "src/rag_agents/retrieval_graph"
"index_graph" = "src/rag_agents/index_graph"
"shared" = "src/rag_agents/shared"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
include = ["*.py", "*.pyi", "*.ipynb"]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]
[tool.ruff.lint.pydocstyle]
convention = "google"
